+++ src/main/java/com/kr1s1s/subtlyd/data/tags/ItemTagsSD.java
@@ -4,12 +4,15 @@ import com.kr1s1s.subtlyd.SubtlyDungeons;
 import com.kr1s1s.subtlyd.world.item.ItemsSD;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricTagProvider;
+import net.minecraft.core.Holder;
 import net.minecraft.core.HolderLookup;
+import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.tags.ItemTags;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.item.Item;
 
+import java.util.List;
 import java.util.concurrent.CompletableFuture;
 
 public class ItemTagsSD extends FabricTagProvider.ItemTagProvider {
@@ -32,5 +35,19 @@ public class ItemTagsSD extends FabricTagProvider.ItemTagProvider {
 
     private static TagKey<Item> create(String string) {
         return TagKey.create(Registries.ITEM, SubtlyDungeons.identifier(string));
+    }
+
+    /**
+     * Can be used to get a list of items by item tag. Cannot be used within data generator classes.
+     * @param tag The specified tag to search.
+     * @return A list of items with the specified item tag.
+     */
+    public static List<Item> getItems(TagKey<Item> tag) {
+        Iterable<Holder<Item>> holders = BuiltInRegistries.ITEM.getTagOrEmpty(tag);
+        List<Item> items = new java.util.ArrayList<>(List.of());
+        for  (Holder<Item> holder : holders) {
+            items.add(holder.value());
+        }
+        return items;
     }
 }